+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * feature: 单元格事件
     - `props.cellHover`启用td hover状态
     - 新增 `cell-mouseenter` `cell-mouseleave` `cell-mouseover` 事件。
+* change: 减少css选择器层级。th background-color 变更为inherit
 * fix: 高亮行时，有概率不触发高亮的问题。
 
 ## 0.3.1